Move Product Collection editor experience components into /edit directory
