+++ assets/js/blocks/product-collection/edit/editor.scss
