+++ assets/js/blocks/product-collection/edit/inspector-controls/hand-picked-products-control.tsx
@@ -16,7 +16,7 @@ import {
 /**
  * Internal dependencies
  */
-import { QueryControlProps } from '../types';
+import { QueryControlProps } from '../../types';
 
 /**
  * Returns:
