+++ assets/js/blocks/product-collection/edit/inspector-controls/on-sale-control.tsx
@@ -12,7 +12,7 @@ import {
 /**
  * Internal dependencies
  */
-import { QueryControlProps } from '../types';
+import { QueryControlProps } from '../../types';
 
 const OnSaleControl = ( props: QueryControlProps ) => {
 	const { query, setQueryAttribute } = props;
