+++ assets/js/blocks/product-collection/edit/inspector-controls/order-by-control.tsx
@@ -16,8 +16,8 @@ import {
 	TProductCollectionOrder,
 	TProductCollectionOrderBy,
 	QueryControlProps,
-} from '../types';
-import { getDefaultQuery } from '../constants';
+} from '../../types';
+import { getDefaultQuery } from '../../constants';
 
 const orderOptions = [
 	{
