+++ assets/js/blocks/product-collection/edit/inspector-controls/taxonomy-controls/taxonomy-item.tsx
