+++ assets/js/blocks/product-collection/edit/inspector-controls/upgrade-notice.tsx
