+++ assets/js/blocks/product-collection/edit/toolbar-controls/display-settings-toolbar.tsx
@@ -16,7 +16,7 @@ import {
 /**
  * Internal dependencies
  */
-import { ProductCollectionQuery } from '../types';
+import { ProductCollectionQuery } from '../../types';
 
 interface DisplaySettingsToolbarProps {
 	query: ProductCollectionQuery;
