+++ assets/js/blocks/product-collection/edit/toolbar-controls/pattern-chooser-toolbar.tsx
