+++ assets/js/blocks/product-collection/edit/toolbar-controls/pattern-selection-modal.tsx
@@ -13,7 +13,7 @@ import { type BlockInstance, cloneBlock } from '@wordpress/blocks';
 /**
  * Internal dependencies
  */
-import { ProductCollectionQuery } from '../types';
+import { ProductCollectionQuery } from '../../types';
 
 const blockName = 'woocommerce/product-collection';
